feat(Earthfile): Add target to build RPM, Deb and APK packages for crank
This commit adds new targets for building Deb, RPM and APK packages for
crank, so that they can be used in popular Linux distributions, along
with a multiplatform-build target for each one of them, so that we build
packages for supported archs on major Linux distributions.

Related ticket https://github.com/crossplane/crossplane/issues/4899

There is also a `+lib-pkg-build` target added, which is supposed to be
used by the other targets to do the actual build of the packages. It
uses nFPM for building the packages, see https://nfpm.goreleaser.com/
The config that nFPM is used is also added in this commit. The distro
package is called `crossplane`.

Additionally, there is a `pkg-build` target added, which calls all the
package format specific targets, so that we can build all of them
locally for our current architecture.

Finally, there is also another target added `+pkg-multiplatform-build`
that builds all the packages for all the supported linux architectures.
This commit also adds this target in the `+multiplatform-build` one, so
that the packages are also built by the CI and uploaded as artifacts.

This is not a complete solution, as the end users will still need to
download manually the package, or build it locally, in order to install
it. This would also not provide automatic updates. Ideally there should
be repositories provided for some of the popular Linux distros. Packages
and repositories should also be signed with a trusted project key.