feat(Earthfile): Add target to build RPM, Deb and APK packages for crank
This commit adds a new target called `+pkg-build` that builds Deb, RPM
and APK packages for crank, so that they can be used in popular linux
distributions.

It uses nFPM for building the packages, see https://nfpm.goreleaser.com/
The config that nFPM is used is also added in this commit. The distro
package is called `crossplane`.

There is also another target `+pkg-multiplatform-build` that builds the
packages for all the supported linux architectures. I am including this
target in the `+multiplatform-build` one, so that the packages are also
built by the CI and uploaded as artifacts to S3.

TODO Improve this section

Testing was performed in the following distributions

Ubuntu $VERSION? :

dpkg -i crossplane.deb

Debian $VERSION? :

Fedora $VERSION? :

openSUSE $VERSION? :

Issues:
- the packages are not signed
- This is not a complete solution, we need to set up repositories so
  that the end users get automatically the upgrades of the packages.